backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
pydantic-settings>=2.10,<3.0
python-multipart>=0.0.20,<1.0
pytest>=8.4,<9.0
httpx>=0.28,<1.0
openai>=1.54,<2.0
psycopg[binary]>=3.2,<4.0
boto3>=1.35,<2.0
chromadb>=0.5,<1.0
